MetadataAccessObject (pure virtual) /\ | | RDBMSMetadataAccessObject (old MetadataAccessObject) <>----QueryExecutor                                                                    /\                                                                     |                                                         QueryConfigExecutor
MetadataSource is still there, but I have left it out because it makes the diagram more complex.

RelationDBMAO would be a final class. The differences between the existing and spanner implementations would be represented in a QueryExecutor class (using composition over inheritance). If, in the future new differences arose, those would be moved to the QueryExecutor class. For example, InitMetadataSource, while not a query per se, would be impemented in QueryExecutor, while "CREATE TABLE", would not be.

PiperOrigin-RevId: 289549228