Adding name for context, execution and artifact
When a client calls the MLMD CREATE APIs, the client might not guarantee to issue request to MLMD API once and only once, in cases such as failover or retry. It’s important for MLMD to provide an idempotency guarantee to avoid creating duplicate dangling nodes.

The requirements above can be achieved by accepting a unique (in the scope of a database ID), client-provided name for creating the resource. The name will be used by clients for CRUD the resources. The existing id will be used by MLMD internally as primary key for tracking and indexing.

PiperOrigin-RevId: 298393926