Allows type upsert methods omitting stored properties.
This is introduced to write forward compatible MLMD type registration routines for the MLMD applications.

The current type registration method enforces the callsite must provide all stored properties when upserting types. For old releases (e.g., tfx), the callsite may not be aware of any newly introduced properties by future releases. The change introduce a mode to loose the constraint and allow the callsite only give a subset properties in the stored types.

PiperOrigin-RevId: 342083117