Introduces SpannerMetadataAccessObjectCreationOptions in metadata_source.proto.
SpannerMetadataAccessObjectCreationOptions is introduced as a preparation for enabling BufferWrite() on MLMD Service with CDPush. This proto type would be defined using CDPush framework and be passed down to CreateSpannerMetadataAccessObject(). It would be used for initializing the `enable_bufferwrite_` flag for SpannerQueryExecutor which controls the Spanner transaction to issue Write() or BufferWrite().

BEGIN_PUBLIC
no-op
END_PUBLIC

PiperOrigin-RevId: 472763612