Makes `enable_bufferwrite` available for MLMD Service through CDPush.
NOTICE: Since this feature uses protoconf.pi to control, which is essentially CDPush, submitting this CL will NOT be effective in ANY env. This feature will be enabled in non-prod env first after the CL is rolled out in staging.

Impact evaluation:
This feature will enable Artifact/Execution/Context, Event, Attribution/Association's BufferWrite(). The insertion performance for Spanner backend accordingly will be optimized by reducing the number of RPCs sent for insertion operations from O(N(*M)) to O(1).  N is the number of entities to be batch inserted and M is the number of properties when inserting Artifact/Execution/Context.

When inserting 10 Artifact/Execution/Context with 3 properties and 1 custom properties, we reduce the mean latency by ~37%, from ~570ms to ~360ms as shown in [1], [2] and [3].
When inserting 10 Events, we reduce the mean latency by ~12.5%, from ~400ms to ~350ms as shown in [4].
When inserting 10 Attributions and Associations, we reduce the mean latency by ~5%, from ~640ms to ~610ms as shown in [5].

The improvement is even larger if we increase the number of inserted entities from 10 to 100: reduce ~60% from ~4100ms to ~1600ms for Artifact/Execution/Context, ~21% from ~1900ms to ~1500ms for Event and ~17% from ~4700ms to ~4000ms for Attributions/Associations as shown in [6].

Moreover, enabling the BufferWrite() does not cause any performance regressions to our existing read load testings as shown in [7].

The impact in prod would be closely monitored in the future.

Next step for optimizing MLMD Service API performance would be to introduce foreign key constraint in MLMD Spanner schema so that we can reduce most of current fan-out `ExecuteQuery` caused by pre-checks [8] - which is also the reason why Event and Attribution/Association do not have a very large improvement by enabling BufferWrite(). For more details, please look at go/mlmd-service-batch-apis-optimization.

[1] https://perfgate.googleplex.com/run?run_key=6236514204516352
[2] https://perfgate.googleplex.com/run?run_key=6081856626491392
[3] https://perfgate.googleplex.com/run?run_key=5563995038416896
[4] https://perfgate.googleplex.com/run?run_key=6210419895828480
[5] https://perfgate.googleplex.com/run?run_key=5535322574946304
[6] http://sponge2/ccb98586-1e7b-4ba8-ab98-b454fb2086d4
[7] http://sponge2/cc035ace-0b40-4332-bf42-e046649fbd84.
[8] https://screenshot.googleplex.com/3Lc6XQ8hZDJ5Qfi

BEGIN_PUBLIC
no-op
END_PUBLIC

PiperOrigin-RevId: 474454663